+++ ascenderkit/api/pages/inventory.py
@@ -142,7 +142,17 @@ class Inventories(page.PageList, Inventory):
     pass
 
 
-page.register_page([resources.inventories, resources.related_inventories, resources.constructed_inventories, resources.federated_inventories], Inventories)
+page.register_page(
+    [
+        resources.inventories,
+        resources.related_inventories,
+        resources.constructed_inventories,
+        resources.federated_inventories,
+        resources.host_related_smart_inventories,
+        resources.inventory_input_inventories,
+    ],
+    Inventories,
+)
 
 
 class Group(HasCreate, HasVariables, base.Base):
@@ -227,6 +237,7 @@ page.register_page(
     [
         resources.groups,
         resources.host_groups,
+        resources.host_related_all_groups,
         resources.inventory_related_groups,
         resources.inventory_related_root_groups,
         resources.group_children,
@@ -384,7 +395,7 @@ class InventorySources(page.PageList, InventorySource):
     pass
 
 
-page.register_page([resources.inventory_sources, resources.related_inventory_sources], InventorySources)
+page.register_page([resources.inventory_sources, resources.related_inventory_sources, resources.project_scm_inventory_sources], InventorySources)
 
 
 class InventorySourceGroups(page.PageList, Group):
@@ -420,6 +431,17 @@ class InventoryUpdateCancel(base.Base):
 
 
 page.register_page(resources.inventory_update_cancel, InventoryUpdateCancel)
+
+
+class InventoryUpdateEvent(base.Base):
+    pass
+
+
+class InventoryUpdateEvents(page.PageList, InventoryUpdateEvent):
+    pass
+
+
+page.register_page(resources.inventory_update_events, InventoryUpdateEvents)
 
 
 class InventoryCopy(base.Base):