+++ ascenderkit/api/pages/workflow_approvals.py
@@ -1,5 +1,6 @@
 from ascenderkit.api.pages import UnifiedJob
 from ascenderkit.api.resources import resources
+from . import base
 from . import page
 from ascenderkit import exceptions
 
@@ -25,4 +26,18 @@ class WorkflowApprovals(page.PageList, WorkflowApproval):
     pass
 
 
-page.register_page(resources.workflow_approvals, WorkflowApprovals)
+page.register_page([resources.workflow_approvals, resources.workflow_approval_template_approvals], WorkflowApprovals)
+
+
+class WorkflowApprovalVote(base.Base):
+    pass
+
+
+page.register_page(resources.workflow_approval_vote, WorkflowApprovalVote)
+
+
+class WorkflowApprovalVotes(page.PageList, WorkflowApprovalVote):
+    pass
+
+
+page.register_page([resources.workflow_approval_votes, resources.workflow_approval_related_votes], WorkflowApprovalVotes)